Fix, Update and add more accessibility.
This adds a horse/mule/donkey/llama status to inventory, enabling the player to view it's current status.

Next, ViaFabric will be updated to 0.4.0+154-main, Fabric-API to 0.34.1 and Alternate-EntityCulling to 1.2.2.

And finally, Advancements will be patched from crashing game when attempting to debug them and README will include it's links.