Update sodium from quick-port to dev build.
This needs a bit of explaining:

The first thing that was discovered is that in the dev build itself; Sodium does in fact NOT support LWJGL 3.2.3 and newer versions beyond it unlike iris' variant.

However the second discovery was that because iris-sodium does support LWJGL 3.2.3 and newer versions beyond it; it actually doesn't know what to do with the memory properly and ends up leaking it in process.

And besides, i do NOT wish to include iris in my repository since my modpack was never designed to support optifine's shaders in any way and instead try to give performance.

Materials that are referencing these issues:

sodium-fabric's issue #748,
sodium-fabric's commit df1efdb,
iris' sodium-fabric fork commit 6eb0140.